[B] Remove google (finally!), rewrite oauth2 in gradesapi
Tons of removed files in this commit. That's because almost all legacy (like, when this was a proxy-- remember that?!) code has been removed.

Google OAuth code also has been removed because, if we are going to add it again, we'll want it to match the newest code structure rather than be old crap.